Fixed Database, product added cshtml and made button work

+++ laohaldusprojekt/laohaldusprojekt/ViewModel/ProductIndexViewModel.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string TooteNimi { get; set; }
         public int Kogus { get; set; }
-        public decimal Hind { get; set; }
+        public double Hind { get; set; }
         public string Kategooria { get; set; }
 
     }
